# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FamilyCreate": {
        "properties": {
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "family_seq": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Seq"
          },
          "family_type": {
            "enum": [
              "patent",
              "trademark",
              "design",
              "advisory"
            ],
            "title": "Family Type",
            "type": "string"
          },
          "restricted": {
            "default": false,
            "title": "Restricted",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tm_design": {
            "default": false,
            "title": "Tm Design",
            "type": "boolean"
          }
        },
        "required": [
          "client_id",
          "title",
          "family_type"
        ],
        "title": "FamilyCreate",
        "type": "object"
      },
      "FamilyOut": {
        "properties": {
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_reference": {
            "title": "Display Reference",
            "type": "string"
          },
          "family_seq": {
            "title": "Family Seq",
            "type": "string"
          },
          "family_type": {
            "enum": [
              "patent",
              "trademark",
              "design",
              "advisory"
            ],
            "title": "Family Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "restricted": {
            "title": "Restricted",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tm_design": {
            "title": "Tm Design",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "client_id",
          "family_seq",
          "reference",
          "display_reference",
          "title",
          "family_type",
          "tm_design",
          "restricted",
          "created_at"
        ],
        "title": "FamilyOut",
        "type": "object"
      },
      "FamilyRecord": {
        "description": "The canonical, versioned Family Record.",
        "properties": {
          "applicant": {
            "description": "Client (applicant/owner) name",
            "title": "Applicant",
            "type": "string"
          },
          "client_code": {
            "title": "Client Code",
            "type": "string"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/CountryRecord"
            },
            "title": "Countries",
            "type": "array"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "family_type": {
            "title": "Family Type",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "reference",
          "title",
          "family_type",
          "client_code",
          "applicant",
          "generated_at"
        ],
        "title": "FamilyRecord",
        "type": "object"
      },
      "FamilyRecordExport": {
        "description": "Signed, shareable snapshot (spec \u00a73.1) \u2014 the controlled foreign-associate sharing surface\ntoday, the Client Portal substrate (D34) and chain-write path later.",
        "properties": {
          "key_id": {
            "description": "Identifier of the signing key used",
            "title": "Key Id",
            "type": "string"
          },
          "record": {
            "$ref": "#/components/schemas/FamilyRecord"
          },
          "signature": {
            "description": "HMAC-SHA256 over the canonical record JSON (v1 interim)",
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "record",
          "signature",
          "key_id"
        ],
        "title": "FamilyRecordExport",
        "type": "object"
      },
      "FamilyUpdate": {
        "properties": {
          "family_type": {
            "anyOf": [
              {
                "enum": [
                  "patent",
                  "trademark",
                  "design",
                  "advisory"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Type"
          },
          "restricted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restricted"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tm_design": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tm Design"
          }
        },
        "title": "FamilyUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterCreate": {
        "properties": {
          "application_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application No"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "filing_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "jurisdiction_code": {
            "title": "Jurisdiction Code",
            "type": "string"
          },
          "parent_matter_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Matter Id"
          },
          "registration_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Date"
          },
          "registration_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration No"
          },
          "relationship_type": {
            "anyOf": [
              {
                "enum": [
                  "continuation",
                  "cip",
                  "divisional",
                  "pct_national_phase",
                  "madrid_designation",
                  "tm_extension",
                  "external_priority",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Type"
          },
          "responsible_associate_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Associate Id"
          },
          "responsible_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible User Id"
          },
          "segment_base": {
            "title": "Segment Base",
            "type": "string"
          },
          "small_entity": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Small Entity"
          },
          "status": {
            "default": "pending",
            "enum": [
              "pending",
              "filed",
              "published",
              "allowed",
              "issued",
              "registered",
              "abandoned",
              "client_abandoned",
              "expired",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "jurisdiction_code",
          "segment_base"
        ],
        "title": "MatterCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "IP practice-management platform API (WP 0.7 scaffold).",
    "title": "BrunetCo OS API",
    "version": "0.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/families": {
      "post": {
        "operationId": "create_family_api_v1_families_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Family",
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/families/{family_id}": {
      "get": {
        "operationId": "get_family_api_v1_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Family",
        "tags": [
          "families"
        ]
      },
      "patch": {
        "operationId": "update_family_api_v1_families__family_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Family",
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/families/{family_id}/export": {
      "get": {
        "operationId": "get_family_export_api_v1_families__family_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyRecordExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Family Export",
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/families/{family_id}/record": {
      "get": {
        "operationId": "get_family_record_api_v1_families__family_id__record_get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Family Record",
        "tags": [
          "families"
        ]
      }
    },
...
    "/api/v1/matters": {
      "post": {
        "operationId": "create_matter_api_v1_matters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Matter",
        "tags": [
          "matters"
        ]
      }
    },
    "/api/v1/matters/{matter_id}": {
      "get": {
        "operationId": "get_matter_api_v1_matters__matter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matter",
        "tags": [
          "matters"
        ]
      },
      "patch": {
        "operationId": "update_matter_api_v1_matters__matter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Matter",
        "tags": [
          "matters"
        ]
      }
    },
...
  }
}